requirements.txt
flask
numpy
requests
rio_tiler==1.1.3
flask_compress
flask_cors
gunicorn
pyproj==2.1.3
pillow
utm
shapely
rasterstats
gevent